[controller][common][compat] Add rollback/rollforward endpoints to support per colo version swap (#893)
The endpoint for rollback/rollforward of current version allows global version swap across all regions and it works only if each region contains the exact same versions. This PR adds support for individual version swap in each colo using admin channel. Also added some schema changes to store config to enable unused schema deletion on a per store basis

---------

+++ services/venice-controller/src/main/java/com/linkedin/venice/controller/UnusedValueSchemaCleanupService.java
@@ -91,7 +91,7 @@ public class UnusedValueSchemaCleanupService extends AbstractVeniceService {
                   clusterName,
                   storeName,
                   schemasToDelete);
-              if (cleanupEnabled) {
+              if (cleanupEnabled && store.isUnusedSchemaDeletionEnabled()) {
                 veniceParentHelixAdmin.deleteValueSchemas(clusterName, store.getName(), schemasToDelete);
               }
             }
